requirements.txt
PyAutoGUI==0.9.53
opencv-python==4.4.0.40
numpy==1.23.5
portalocker==2.6.0
torch==1.12.1
PyYAML==6.0
tqdm==4.64.1
matplotlib==3.6.2
pandas==1.5.2
seaborn==0.12.2
Pillow==9.3.0
torchvision==0.13.1
requests==2.28.1
Flask==2.1.3
thop==0.1.1.post2209072238
tensorflow==2.10.1
onnxruntime==1.13.1
rapidfuzz==2.13.7
cnocr==2.2.2.1
pywin32==305
colorlog==6.7.0
scikit-image==0.19.3
jieba==0.42.1
scipy==1.10.0